Separate /profile into subfiles
