+++ src/services/profile/index.ts
@@ -1,20 +1,32 @@
 export {
   getProfile,
   updateProfile,
-  getTFAToken,
-  disableTwoFactor,
-  confirmTwoFactor,
-  createAppToken,
-  getAppToken,
-  updateAppToken,
-  deleteAppToken,
-  getAppTokens,
+} from './profile';
+
+export {
   createPersonalAccessToken,
   getPersonalAccessToken,
   updatePersonalAccessToken,
   deletePersonalAccessToken,
   getPersonalAccessTokens,
+} from './accessTokens';
+
+export {
+  createAppToken,
+  getAppToken,
+  updateAppToken,
+  deleteAppToken,
+  getAppTokens,
+} from './appTokens';
+
+export {
   getSSHKeys,
   createSSHKey,
   deleteSSHKey
-} from './profile';
+} from './sshkeys';
+
+export {
+  getTFAToken,
+  disableTwoFactor,
+  confirmTwoFactor,
+} from './twoFactor';
