+++ src/services/profile/twoFactor.ts
@@ -1,0 +1,46 @@
+import { API_ROOT } from 'src/constants';
+
+import Request, { setData, setMethod, setURL } from '..';
+
+type Secret = Linode.Secret;
+
+/**
+ * getTFAToken
+ *
+ * Generate a token for enabling two-factor authentication.
+ * Used for authorizing 3rd party apps such as Authy and 
+ * Google Authenticator.
+ * 
+ * @returns { Promise }
+ * 
+ * @example updateProfile({"email_notifications": "true"});
+ */
+export const getTFAToken = () =>
+  Request<Secret>(
+    setMethod('POST'),
+    setURL(`${API_ROOT}/profile/tfa-enable`)
+  )
+
+/**
+ * disableTwoFactor
+ *
+ * Disable two-factor authentication for the current user.
+ * All tokens generated by authorized apps will no longer
+ * be valid.
+ * 
+ * @returns { Promise }
+ * 
+ * @example disableTwoFactor();
+ */
+export const disableTwoFactor = () =>
+  Request<{}>(
+    setMethod('POST'),
+    setURL(`${API_ROOT}/profile/tfa-disable`)
+  )
+
+export const confirmTwoFactor = (code: string) =>
+  Request<{}>(
+    setMethod('POST'),
+    setURL(`${API_ROOT}/profile/tfa-enable-confirm`),
+    setData({ tfa_code: code })
+  )
